Create convoluzionali.py
https://www.aiutocomputerhelp.it/machine-learning-capitolo-3-reti-neurali-convoluzionali-software/

Convolutional Neural Networks (CNN), also known as Convolutional Neural Networks, are one of the fundamental pillars of Deep Learning, especially in the field of computer vision. First introduced in the 1980s and made famous by the successes of models such as LeNet and AlexNet, CNNs have revolutionized the way artificial intelligence systems perceive and understand images.